vehicle_mechanic_coach/requirements.txt
streamlit>=1.32
groq>=0.5
markdown>=3.5
python-dotenv>=1.0
sentence-transformers>=2.7
faiss-cpu>=1.8
pyyaml>=6.0
pdfplumber==0.11.7
pdfminer.six==20250324
pillow>=10.0
pypdfium2>=4.30.0


